Брахиозавр

+++ dino.h
@@ -158,7 +158,7 @@ struct Dino
     std::multiset<double, std::greater<double>> shield{0};
     std::multiset<std::pair<double, double>, dodge_cmp> dodge{std::make_pair(0, 0)};
     int cooldown[4] = {};
-    int prepared_damage_factor = 0;
+    double prepared_damage_factor = 0;
     bool crit = false;
     bool killer = false;
